+++ packages/workshop-cli/tsconfig.build.json
@@ -1,4 +1,5 @@
 {
+	"extends": "./tsconfig.json",
 	"compilerOptions": {
 		"declaration": true,
 		"declarationMap": true,
@@ -6,13 +7,17 @@
 		"forceConsistentCasingInFileNames": true,
 		"inlineSources": true,
 		"jsx": "react",
-		"module": "nodenext",
-		"moduleResolution": "nodenext",
+		"module": "esnext",
+		"moduleResolution": "bundler",
 		"noUncheckedIndexedAccess": true,
 		"resolveJsonModule": true,
 		"skipLibCheck": true,
 		"sourceMap": true,
 		"strict": true,
-		"target": "es2022"
-	}
+		"target": "es2022",
+		"outDir": "./dist",
+		"rootDir": "./src"
+	},
+	"include": ["src/**/*"],
+	"exclude": ["node_modules", "dist"]
 }
